Ticket #488 Add file size, file metadata and job metadata to fts-transfer-submit CLI

+++ src/db/schema/mysql-schema.sql
@@ -559,7 +559,7 @@ CREATE TABLE t_file (
   
 --
 -- user provided size of the file (bytes)
-  user_filesize  BIGINT,  
+  user_filesize  INTEGER,  
   
 --
 -- File metadata
